assignUsersToTemplate
