+++ lib/templates.ts
@@ -7,132 +7,12 @@ import { AuditLogAccessDeniedDetails, AuditLogDetails, AuditLogEvent, logEvent }
 import { logMessage } from "@lib/logger";
 import { unprocessedSubmissions } from "./vault";
 import { deleteKey } from "./serviceAccount";
-import { sendEmail } from "./integration/notifyConnector";
-import { ownerAddedEmailTemplate } from "./invitations/emailTemplates/ownerAddedEmailTemplate";
+
 import { isValidISODate } from "./utils/date/isValidISODate";
 import { validateTemplate } from "@lib/utils/form-builder/validate";
 import { dateHasPast } from "@lib/utils";
 import { validateTemplateSize } from "@lib/utils/validateTemplateSize";
 import { invalidateTemplateEditLockUserCountCache } from "./editLocks";
-
-/**
- * Assign a user to a form
- *
- * @param ability
- * @param formID
- * @param userID
- */
-export async function assignUserToTemplate(formID: string, userID: string): Promise<void> {
-  const { user } = await authorization.canEditForm(formID).catch((e) => {
-    logEvent(
-      e.user.id,
-      { type: "Form", id: formID },
-      "AccessDenied",
-      AuditLogAccessDeniedDetails.AccessDenied_AttemptToRemoveAssignedUser
-    );
-    throw e;
-  });
-
-  const template = await prisma.template.findUnique({
-    where: {
-      id: formID,
-    },
-    select: {
-      users: {
-        select: {
-          id: true,
-          name: true,
-          email: true,
-        },
-      },
-    },
-  });
-
-  if (template === null) {
-    logMessage.warn(`Can not add user ${userID} to template ${formID}.  Template does not exist`);
-    throw new TemplateNotFoundError();
-  }
-
-  const userToAdd = template.users.find((user) => user.id === userID);
-
-  if (!userToAdd) {
-    logMessage.warn(`Can not add user ${userID} to template ${formID}.  User does not exist`);
-    throw new UserNotFoundError();
-  }
-
-  const updatedTemplate = await prisma.template
-    .update({
-      where: {
-        id: formID,
-      },
-      select: {
-        jsonConfig: true,
-        users: {
-          select: {
-            id: true,
-            email: true,
-          },
-        },
-      },
-      data: {
-        users: {
-          connect: {
-            id: userID,
-          },
-        },
-      },
-    })
-    .catch((e) => prismaErrors(e, null));
-
-  // No changes
-  if (updatedTemplate === null) return;
-
-  await invalidateTemplateEditLockUserCountCache(formID);
-
-  logEvent(
-    user.id,
-    { type: "Form", id: formID },
-    AuditLogEvent.InvitationAccepted,
-    AuditLogDetails.AcceptedInvitation,
-    { userEmail: user.email }
-  );
-
-  notifyOwnersOwnerAdded(
-    userToAdd,
-    updatedTemplate.jsonConfig as FormProperties,
-    updatedTemplate.users
-  );
-}
-
-/**
- * Notify all owners when ownership changes (owner added)
- *
- * @param user New owner
- * @param form Form properties object
- * @param users Current owners
- */
-export const notifyOwnersOwnerAdded = async (
-  userToAdd: { name: string | null; email: string },
-  form: FormProperties,
-  users: { id: string; email: string }[]
-) => {
-  const emailContent = ownerAddedEmailTemplate(
-    form.titleEn,
-    form.titleFr,
-    userToAdd.name || userToAdd.email
-  );
-
-  users.forEach((owner) => {
-    sendEmail(
-      owner.email,
-      {
-        subject: "Ownership change notification | Notification de changement de propriété",
-        formResponse: emailContent,
-      },
-      "notifyAddedOwner"
-    );
-  });
-};
 
 /**
  * Add/remove (sync) users to a form
